Make it possible to show checkpoints in source control panel
Had to duplicate a lot of config in package.json because the
evaluation of "when" clauses is limited.
It is only possible to use configuration (prefix config.) in
"when" clauses if it evaluates to a boolean (forces us to use two
configs for showCheckpointsIn*).
The new location of the tree view requires a new view ID,
which in turn requires us to duplicate all menu item commands
( "||" doesn't work in when clause).